resolved issue 179
