+++ skunkworks_crow/src/main/java/org/odk/share/injection/FragmentBuilder.java
@@ -3,7 +3,6 @@ package org.odk.share.injection;
 import org.odk.share.views.ui.send.fragment.BlankFormsFragment;
 import org.odk.share.views.ui.send.fragment.FilledFormsFragment;
 import org.odk.share.views.ui.instance.fragment.ReceivedInstancesFragment;
-import org.odk.share.views.ui.instance.fragment.ReviewedInstancesFragment;
 import org.odk.share.views.ui.instance.fragment.SentInstancesFragment;
 import org.odk.share.views.ui.instance.fragment.StatisticsFragment;
 import org.odk.share.injection.config.scopes.PerActivity;
@@ -25,10 +24,6 @@ public abstract class FragmentBuilder {
     @PerActivity
     @ContributesAndroidInjector
     abstract ReceivedInstancesFragment provideReceivedInstancesFragment();
-
-    @PerActivity
-    @ContributesAndroidInjector
-    abstract ReviewedInstancesFragment provideReviewedInstancesFragment();
 
     @PerActivity
     @ContributesAndroidInjector
