+++ skunkworks_crow/src/main/java/org/odk/share/services/SenderService.java
@@ -19,7 +19,7 @@ import javax.inject.Singleton;
 
 import timber.log.Timber;
 
-import static org.odk.share.views.ui.instance.fragment.ReviewedInstancesFragment.MODE;
+import static org.odk.share.views.ui.instance.fragment.ReceivedInstancesFragment.MODE;
 
 @Singleton
 public class SenderService {
