+++ skunkworks_crow/src/main/res/layout/list_item_checkbox.xml
@@ -1,5 +1,6 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/parent"
     style="@style/MatchParentWidth"
     android:background="?attr/selectableItemBackground">
