+++ skunkworks_crow/src/main/res/values/colors.xml
@@ -5,4 +5,5 @@
     <color name="colorAccent">#00796B</color>
     <color name="colorTabInactive">#CFD8DC</color>
     <color name="colorTabActive">#FFFFFF</color>
+    <color name="colorSelected">#50909090</color>
 </resources>
