+++ skunkworks_crow/src/main/res/values/strings.xml
@@ -12,6 +12,7 @@
     <string name="saved_forms">Saved Forms</string>
     <string name="send_selected">Send</string>
     <string name="select_all">Select all</string>
+    <string name="send_reviewed_forms">Send reviewed forms</string>
     <string name="clear_all">Clear all</string>
     <string name="send_count">Send (%d)</string>
     <string name="send_forms">Send</string>
